+++ DEDA_class2019_SYSU_Abstract_LDA_Crawler/metainfo.txt
@@ -1,0 +1,9 @@
+Name of Quantlet: DEDA_class2019_SYSU_Abstract_LDA
+
+Published in:  LN_SDA_2019
+
+Description: Collect 130 abstracts from http://www.wiwi.hu-berlin.de/de/forschung/irtg/results/discussion-papers
+
+Keywords: web scraper, regular expression, abstract analysis
+
+Author: FB, FS
