--- LDA_demo/README.md
@@ -1,25 +1,0 @@
-[<img src="https://github.com/QuantLet/Styleguide-and-FAQ/blob/master/pictures/banner.png" width="888" alt="Visit QuantNet">](http://quantlet.de/)
-
-## [<img src="https://github.com/QuantLet/Styleguide-and-FAQ/blob/master/pictures/qloqo.png" alt="Visit QuantNet">](http://quantlet.de/) **FB_2019_Abstract_Analysis** [<img src="https://github.com/QuantLet/Styleguide-and-FAQ/blob/master/pictures/QN2.png" width="60" alt="Visit QuantNet 2.0">](http://quantlet.de/)
-
-```yaml
-
-Name of Quantlet: FB_2019_Abstract_Analysis
-
-Published in:  LN_SDA_2019
-
-Description: Collect 130 abstracts and create a word cloud picture to make a short description of these 130 abstracts; Summarize 5 topics from 60 abstracts with LDA method
-
-Keywords: LDA, word cloud, topic modelling, abstract analysis
-
-See also: LDA-DTM_Christmas Songs, SL_2019_alpha_centauri_pirates_wordcloud
-
-Author: FB, FS
-
-```
-
-![Picture1](heatmap_abstract.png)
-
-![Picture2](qletlogo_tr.png)
-
-![Picture3](wordcloud_abstract.png)
